+++ src/qibolab/platform.py
@@ -8,7 +8,7 @@ from typing import Dict, List, Optional
 import networkx as nx
 from qibo.config import log, raise_error
 
-from qibolab.couplers import Coupler, CouplerId, CouplerPair, CouplerPairId
+from qibolab.couplers import Coupler, CouplerId
 from qibolab.execution_parameters import ExecutionParameters
 from qibolab.instruments.abstract import Controller, Instrument, InstrumentId
 from qibolab.native import NativeType
@@ -20,7 +20,6 @@ InstrumentMap = Dict[InstrumentId, Instrument]
 QubitMap = Dict[QubitId, Qubit]
 CouplerMap = Dict[CouplerId, Coupler]
 QubitPairMap = Dict[QubitPairId, QubitPair]
-CouplerPairMap = Dict[CouplerPairId, CouplerPair]
 
 
 @dataclass
@@ -57,8 +56,6 @@ class Platform:
 
     couplers: Optional[CouplerMap] = None
     """Dictionary mapping coupler names to :class:`qibolab.couplers.Coupler` objects."""
-    coupler_pairs: Optional[CouplerPairMap] = None
-    """Dictionary mapping sorted tuples of coupler names to :class:`qibolab.qubits.QubitPair` objects."""
 
     is_connected: bool = False
     """Flag for whether we are connected to the physical instruments."""
